test(inference): extract smoke verifier harness
